requirements.txt
asgiref==3.4.1
attrs==22.2.0
autobahn==21.2.1
Automat==22.10.0
cffi==1.15.1
channels==3.0.4
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.2
Django==3.2.16
djangorestframework==3.12.1

hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pytz==2022.7.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
zope.interface==5.5.2
